requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23
opencv-python>=4.8
albumentations>=1.3
pyyaml>=6.0
tqdm>=4.64
tensorboard>=2.12
matplotlib>=3.7
pillow>=9.5
